metier affectation trial

+++ src/main/java/com/example/applicationmaven1/repositories/FoyerRepository.java
@@ -3,5 +3,8 @@ package com.example.applicationmaven1.repositories;
 import com.example.applicationmaven1.Entities.Foyer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FoyerRepository extends JpaRepository<Foyer, Long> {
+import java.util.Optional;
+
+public interface FoyerRepository extends JpaRepository<Foyer, Integer> {
+
 }
